Work on the iPad viewer

+++ iPad Viewer/Model/LoggerTCPTransport.h
@@ -45,11 +45,11 @@
 	CFSocketRef			listenerSocket_ipv4;
 	CFSocketRef			listenerSocket_ipv6;
 	NSString			*bonjourServiceName;
-	int					listenerPort;
+	NSInteger			listenerPort;
 	BOOL				publishBonjourService;
 }
 
-@property (nonatomic, assign) int listenerPort;
+@property (nonatomic, assign) NSInteger listenerPort;
 @property (nonatomic, assign) BOOL publishBonjourService;
 @property (nonatomic, readonly) CFSocketRef listenerSocket_ipv4;
 @property (nonatomic, readonly) CFSocketRef listenerSocket_ipv6;
